[375] 최소수수료율 0.1 지원

+++ lib/providers/node_provider/isolate/isolate_manager.dart
@@ -440,7 +440,7 @@ class IsolateManager {
     return _sendWithSocketCheck(IsolateControllerCommand.broadcast, [signedTx]);
   }
 
-  Future<Result<int>> getNetworkMinimumFeeRate() async {
+  Future<Result<double>> getNetworkMinimumFeeRate() async {
     return _sendWithSocketCheck(IsolateControllerCommand.getNetworkMinimumFeeRate, []);
   }
